Group related components into under one file
