+++ src/components/Kanban/Lane/styles.scss
@@ -1,0 +1,6 @@
+@import '../../../assets/styles/_variables.scss';
+@import '../../../assets/styles/_mixins.scss';
+
+.lane {
+  background-color: #ccc;
+}
