+++ src/components/Kanban/LaneHeader/LaneHeaderText/styles.scss
@@ -1,5 +1,5 @@
-@import '../../../assets/styles/_variables.scss';
-@import '../../../assets/styles/_mixins.scss';
+@import '../../../../assets/styles/_variables.scss';
+@import '../../../../assets/styles/_mixins.scss';
 
 .laneHeaderText {
   overflow: hidden;
