+++ src/components/Kanban/LaneHeader/styles.scss
@@ -1,5 +1,5 @@
-@import '../../assets/styles/_variables.scss';
-@import '../../assets/styles/_mixins.scss';
+@import '../../../assets/styles/_variables.scss';
+@import '../../../assets/styles/_mixins.scss';
 
 .laneHeader {
   max-width: 15rem;
