+++ src/components/Kanban/Lanes/styles.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/_variables.scss';
+@import '../../../assets/styles/_variables.scss';
 
 .lanes {
   display: flex;
