+++ src/components/Kanban/Notes/styles.scss
@@ -1,5 +1,5 @@
-@import '../../assets/styles/_variables.scss';
-@import '../../assets/styles/_mixins.scss';
+@import '../../../assets/styles/_variables.scss';
+@import '../../../assets/styles/_mixins.scss';
 
 .notes {
   padding: 1rem;
