--- src/components/Lane/styles.scss
@@ -1,6 +1,0 @@
-@import '../../assets/styles/_variables.scss';
-@import '../../assets/styles/_mixins.scss';
-
-.lane {
-  background-color: #ccc;
-}
